review comments

+++ include/processor.h
@@ -5,7 +5,7 @@
 
 class Processor {
  public:
-  float Utilization();
+  double Utilization();
 
   // private members
  private:
